Salve Bede!!!

+++ src/Conexao4.Application.Contracts/Produtos/ProdutoDto.cs
@@ -9,6 +9,6 @@ namespace Conexao4.Produtos
     {
         public string Nome { get; set; } = string.Empty;
 
-        public float Preco { get; set; }
+        public decimal Preco { get; set; }
     }
 }
